requirements.txt
Django==4.0.1
asgiref==3.4.1
backports.zoneinfo==0.2.1
pip==21.3
setuptools==58.2.0
sqlparse==0.4.2
wheel==0.37.1
djangorestframework==3.13.1
psycopg2==2.9.3
PyJWT==2.3.0
django-environ==0.8.1